+++ src/main/java/com/projarq/dominio/RepCep.java
@@ -1,0 +1,7 @@
+package com.projarq.dominio;
+
+import java.io.IOException;
+
+public interface RepCep {
+    boolean verificaCEP(String cep, String nomeCidade) throws IOException;
+}
