+++ src/main/java/com/projarq/dominio/RepCidade.java
@@ -1,0 +1,15 @@
+package com.projarq.dominio;
+
+import java.util.List;
+
+import com.projarq.dominio.entidades.Cidade;
+
+public interface RepCidade {
+    Cidade salvar(Cidade cidade);
+
+    Cidade findById(Long id);
+
+    List<Cidade> findAll();
+
+    void delete(Long id);
+}
